Remove CancelableTask, minor code cleanup
CancelableTask was overriden entirely by the AsyncTask, so
its implementation wasn't being used at all. Since the nuget
is AsyncTask, for now we'll just keep it to just that task :).

A generic Cancelable task might best be provided by MSBuild
itself as a base class, so maybe we should PR to that repo
instead?

Removed OutputLine type which wasn't being used anywhere and
shortened the code to leverage expression-bodied members to
make the code a bit more readable.